Introduce interpolation options.
When interpolation is enabled, the window will be resizable.
The window size and the scale multiplier are then independent.
The selected interpolation method (nearest neighbour, bilinear, bicubic)
will be used to stretch the image to the window size.

When the "Keep aspect ratio" is checked, the gameboy aspect ratio
will be retained, but the image will not cover the entire window both
horizontally and vertically.

There is still some work to do, like saving these setting to the config
file, and introducing shortcuts for these options.